Move save-button and user
