+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/presentation/workspace-item/workspace-item.tsx
@@ -15,7 +15,7 @@
 import * as React from 'react'
 import styled from 'styled-components'
 import Card from '../card'
-import SaveButton from '../save-button'
+import SaveButton from '../../save-button'
 import { buttonTypeEnum, Button } from '../button'
 import Dropdown from '../dropdown'
 import NavigationBehavior from '../navigation-behavior'
