+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/presentation/workspace-menu/workspace-menu.tsx
@@ -18,7 +18,7 @@ import { hot } from 'react-hot-loader'
 import WorkspaceInteractions from '../../container/workspace-interactions'
 const QueryAddView = require('../../../component/query-add/query-add.view.js')
 import MarionetteRegionContainer from '../../container/marionette-region-container'
-import SaveButton from '../save-button'
+import SaveButton from '../../save-button'
 import WorkspaceTitle from '../workspace-title'
 import Dropdown from '../dropdown'
 import NavigationBehavior from '../navigation-behavior'
