+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/save-button/save-button.tsx
@@ -14,7 +14,7 @@
  **/
 import * as React from 'react'
 import styled from 'styled-components'
-import { Button, buttonTypeEnum } from '../button'
+import { Button, buttonTypeEnum } from '../presentation/button'
 import { hot } from 'react-hot-loader'
 
 type SaveButtonProps = {
