+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/user/container.tsx
@@ -13,8 +13,10 @@
  *
  **/
 import * as React from 'react'
-import UserComponent from '../../presentation/user'
-const user = require('../../../component/singletons/user-instance.js')
+import UserComponent from './presentation'
+import fetch from '../utils/fetch'
+import processActions from '@connexta/atlas/atoms/logout'
+const user = require('../../component/singletons/user-instance.js')
 
 interface State {
   username: string
